# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "Kiwoom Stock API", "version": "1.0" },
  "paths": {
...
    "/api/trading/orders": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TradingOrder" }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PaperOrderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TradingOrder" }
              }
            }
          }
        }
      }
    },
    "/api/trading/orders/{id}/cancel": {
      "post": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TradingOrder" }
              }
            }
          }
        }
      }
    },
...
    "/api/trading/kill-switch": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/KillSwitchRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaperRiskStatus" }
              }
            }
          }
        }
      }
    },
    "/api/trading/kill-switch/resume": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchResumeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaperRiskStatus" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaperOrderRequest": {
        "type": "object",
        "required": ["decisionId", "code", "side", "quantity", "price"],
        "properties": {
          "decisionId": { "type": "string" },
          "code": { "type": "string", "pattern": "^[0-9]{6}$" },
          "side": { "type": "string", "enum": ["BUY", "SELL"] },
          "quantity": { "type": "integer", "format": "int64", "minimum": 1 },
          "price": { "type": "number", "exclusiveMinimum": true, "minimum": 0 }
        }
      },
      "TradingOrder": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "decisionId": { "type": "string" },
          "mode": {
            "type": "string",
            "enum": ["SIGNAL_ONLY", "PAPER", "LIVE"]
          },
          "code": { "type": "string" },
          "side": { "type": "string", "enum": ["BUY", "SELL"] },
          "requestedQuantity": { "type": "integer", "format": "int64" },
          "requestedPrice": { "type": "number" },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "SUBMITTED",
              "ACKNOWLEDGED",
              "PARTIALLY_FILLED",
              "FILLED",
              "CANCELED",
              "REJECTED"
            ]
          },
          "filledQuantity": { "type": "integer", "format": "int64" },
          "averageFillPrice": { "type": "number", "nullable": true },
          "rejectionReason": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "PaperRiskStatus": {
        "type": "object",
        "properties": {
          "equity": { "type": "number" },
          "cash": { "type": "number" },
          "grossExposure": { "type": "number" },
          "openPositionCount": { "type": "integer" },
          "dailyReturnRate": { "type": "number" },
          "drawdownRate": { "type": "number" },
          "maxPositionRate": { "type": "number" },
          "maxGrossExposureRate": { "type": "number" },
          "maxDailyLossRate": { "type": "number" },
          "maxDrawdownRate": { "type": "number" },
          "maxOpenPositions": { "type": "integer" },
          "killSwitchActive": { "type": "boolean" },
          "killSwitchReason": { "type": "string", "nullable": true },
          "checkedAt": { "type": "string", "format": "date-time" }
        }
      },
      "KillSwitchRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": { "reason": { "type": "string" } }
      },
      "KillSwitchResumeRequest": {
        "type": "object",
        "required": ["confirmation", "reason"],
        "properties": {
          "confirmation": { "type": "string" },
          "reason": { "type": "string" }
        }
      },
...
    },
...
  }
}